@@ -26,9 +26,9 @@ class Tag < ActiveRecord::Base
   end
 
   validates :name, presence: :true
-  validates :name, format: { with: /^[\w\.:-]*$/, message: 'can only include letters, numbers, and dashes' }
-  # validates :name, :uniqueness => { case_sensitive: false }
-
+  validates :name, format: { with: /^[\w\.:-]*$/, multiline: true, message: 'can only include letters, numbers, and dashes' }
+  # validates :name, :uniqueness => { case_sensitive: false  }
+                                                            
   def id
     tid
   end
@@ -58,23 +58,23 @@ class Tag < ActiveRecord::Base
         .where('term_data.name = ?', tagname)
         .order('node.views DESC')
         .limit(limit)
-        .includes(:node_tag, :tag)
+        .references(:node_tag, :tag)
   end
 
   # finds recent nodes - should drop "limit" and allow use of chainable .limit()
   def self.find_nodes_by_type(tagnames, type = 'note', limit = 10)
     nodes = Node.where(status: 1, type: type)
-                .includes(:tag)
+                .references(:tag)
                 .where('term_data.name IN (?)', tagnames)
                 #.select(%i[node.nid node.status node.type community_tags.nid community_tags.tid term_data.name term_data.tid])
                 # above select could be added later for further optimization
     # .where('term_data.name IN (?) OR term_data.parent in (?)', tagnames, tagnames) # greedily fetch children
     tags = Tag.where('term_data.name IN (?)', tagnames)
     parents = Node.where(status: 1, type: type)
-                  .includes(:tag)
+                  .references(:tag)
                   .where('term_data.name IN (?)', tags.collect(&:parent))
     Node.where('node.nid IN (?)', (nodes + parents).collect(&:nid))
-        .includes(:revision, :tag)
+        .references(:revision, :tag)
         .where(status: 1)
         .order('node_revisions.timestamp DESC')
         .limit(limit)
@@ -96,7 +96,7 @@ class Tag < ActiveRecord::Base
       tag = Tag.where(name: tagname).last
       next unless tag
       parents = Node.where(status: 1, type: type)
-                    .includes(:revision, :tag)
+                    .references(:revision, :tag)
                     .where('term_data.name LIKE ?', tag.parent)
       nids += tag_nids + parents.collect(&:nid)
     end
@@ -110,7 +110,7 @@ class Tag < ActiveRecord::Base
         .where('term_data.name = ? AND node.views > (?)', tagname, views)
         .order('node.nid DESC')
         .limit(limit)
-        .includes(:node_tag, :tag)
+        .references(:node_tag, :tag)
   end
 
   def self.exists?(tagname, nid)
@@ -192,7 +192,7 @@ class Tag < ActiveRecord::Base
 
   def self.find_research_notes(tagnames, limit = 10)
     Node.research_notes.where(status: 1)
-        .includes(:revision, :tag)
+        .references(:revision, :tag)
         .where('term_data.name IN (?)', tagnames)
         .order('node_revisions.timestamp DESC')
         .limit(limit)
