@@ -101,7 +101,7 @@ class DrupalUsers < ActiveRecord::Base
   end
 
   def liked_notes
-    Node.includes(:node_selections)
+    Node.references(:node_selections)
         .where("type = 'note' AND node_selections.liking = ? AND node_selections.user_id = ? AND node.status = 1", true, uid)
         .order('node_selections.nid DESC')
   end
