@@ -39,7 +39,7 @@ class SearchService
   end
 
   def find_tags(input, limit = 5)
-    Tag.includes(:node)
+    Tag.references(:node)
        .where('node.status = 1')
        .limit(limit)
        .where('name LIKE ?', '%' + input + '%')
